feat(skore-io/keycloak): add GetUser decorator
