+++ packages/keycloak/src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './keycloak.utils'
+export * from './request.helper'
